+++ tests/unit/test_monkey.py
@@ -4,14 +4,14 @@ from django.core.urlresolvers import reverse
 
 def test_patched_url_reverse(settings):
     settings.SITE_URL = ''
-    url = reverse("pages:homepage")
+    url = reverse("page-home")
     assert url == '/'
 
     settings.SITE_URL = 'https://mysite.com/junction'
 
-    url = reverse("pages:homepage")
+    url = reverse("page-home")
     assert url == 'https://mysite.com/junction/'
 
     settings.SITE_URL = '/junction'
-    url = reverse("pages:homepage")
+    url = reverse("page-home")
     assert url == '/junction/'